+++ proofs/compiler/compiler_proof.v
@@ -237,33 +237,31 @@ Proof.
   exact: (List_Forall2_refl _ value_uincl_refl).
 Qed.
 
-Lemma check_removereturnP entries remove_return b :
-  check_removereturn entries remove_return = ok b →
-  ∀ fn, fn \in entries → remove_return fn = None.
-Proof.
-  move => /assertP /eqP h fn /(in_pmap remove_return).
-  case: (remove_return fn) => // r.
-  by rewrite h.
-Qed.
-
-Lemma compiler_third_partP entries (p p' : @sprog _pd _ _asmop) :
-  compiler_third_part aparams cparams entries p = ok p' →
+Lemma compiler_third_partP returned_params (p p' : @sprog _pd _ _asmop) :
+  compiler_third_part aparams cparams returned_params p = ok p' →
   [/\
     ∀ fn (gd: pointer) scs m va scs' m' vr,
-      fn \in entries →
       sem_call (dc:= direct_c) p gd scs m fn va scs' m' vr →
       exists2 vr',
-      List.Forall2 value_uincl vr vr' &
+      let rminfo fn :=
+        match returned_params fn with
+        | Some l =>
+          let l' := map (fun i => if i is None then true else false) l in
+          if all (fun b => b) l' then None else Some l' (* do we want that? *)
+        | None => removereturn cparams p fn
+        end
+      in
+      List.Forall2 value_uincl (fn_keep_only rminfo fn vr) vr' &
       sem_call (dc:= direct_c) p' gd scs m fn va scs' m' vr' &
     ∀ fn m,
       alloc_ok p' fn m → alloc_ok p fn m
   ].
 Proof.
-  rewrite /compiler_third_part; t_xrbindP=> /check_removereturnP ok_rr pa ok_pa.
+  rewrite /compiler_third_part; t_xrbindP=> pa ok_pa.
   rewrite !print_sprogP => ok_pb pc ok_pc.
   rewrite print_sprogP => <- {p'}.
   split.
-  + move => fn gd scs m va scs' m' vr ok_fn exec_p.
+  + move => fn gd scs m va scs' m' vr exec_p.
     have va_refl : List.Forall2 value_uincl va va.
     - exact: List_Forall2_refl.
     apply: compose_pass_uincl.
@@ -276,19 +274,18 @@ Proof.
     rewrite surj_prog.
     have! [vr' [exec_pa]] :=
       (dead_code_tokeep_callPs (hap_is_move_opP haparams) ok_pa va_refl exec_p).
-    rewrite /fn_keep_only (ok_rr _ ok_fn) => vr_vr'.
     by exists vr'.
   rewrite /alloc_ok => fn m alloc_pc fd get_fd.
-  have! [fda ok_fda get_fda] :=
-    (dead_code_prog_tokeep_get_fundef ok_pa get_fd).
+  have [fda ok_fda get_fda] := [elaborate
+    (dead_code_prog_tokeep_get_fundef ok_pa get_fd)].
   have [fdb [get_fdb ok_fdb]] :=
     allocation_proof.all_checked (sip := sip_of_asm_e) ok_pb get_fda.
-  have! [fdc ok_fdc get_fdc] :=
-    (dead_code_prog_tokeep_get_fundef ok_pc get_fdb).
+  have [fdc ok_fdc get_fdc] := [elaborate
+    (dead_code_prog_tokeep_get_fundef ok_pc get_fdb)].
   move: (alloc_pc _ get_fdc).
   have [_ _ ->]:= dead_code_fd_meta ok_fdc.
   rewrite /sf_total_stack.
-  have [ <- <- <- ] := @check_fundef_meta _ _ _ _ _ _ _ (_, fda) _ _ _ ok_fdb.
+  have [ <- <- <- ] := [elaborate @check_fundef_meta _ _ _ _ _ _ _ (_, fda) _ _ _ ok_fdb].
   have [_ _ ->]:= dead_code_fd_meta ok_fda.
   done.
 Qed.
@@ -298,11 +295,38 @@ Lemma compiler_third_part_meta entries (p p' : sprog) :
   p_extra p' = p_extra p.
 Proof.
   rewrite /compiler_third_part.
-  t_xrbindP => _ pa hpa _ pb hpb.
+  t_xrbindP => pa hpa _ pb hpb.
   have! [_ ok_pa] := (dead_code_prog_tokeep_meta hpa).
   have! [] := (dead_code_prog_tokeep_meta hpb).
   rewrite !print_sprogP /= => _ ok_pb <- {p'}.
   by rewrite ok_pb ok_pa.
+Qed.
+
+Lemma compiler_third_part_invariants entries p p' :
+  compiler_third_part aparams cparams entries p = ok p' ->
+  forall fn fd, get_fundef p.(p_funcs) fn = Some fd ->
+  exists fd', get_fundef p'.(p_funcs) fn = Some fd' /\
+  fd.(f_extra).(sf_align_args) = fd'.(f_extra).(sf_align_args).
+Proof.
+  rewrite /compiler_third_part.
+  t_xrbindP => pa hpa.
+  rewrite 2!print_sprogP /= => check_pa pb hpb.
+  rewrite print_sprogP => <-{p'}.
+  move=> fn fd ok_fd.
+  have [fda ok_fda get_fda] := [elaborate
+    dead_code_prog_tokeep_get_fundef hpa ok_fd].
+  rewrite /check_sprog in check_pa.
+  have [fda' [get_fda' check_fda']] := [elaborate
+    allocation_proof.all_checked (pT:=progStack)
+      (p2:={| p_funcs := regalloc cparams (p_funcs pa); p_globs := p_globs pa; p_extra := p_extra pa |})
+      check_pa get_fda].
+  have [fdb ok_fdb get_fdb] := [elaborate
+    dead_code_prog_tokeep_get_fundef hpb get_fda'].
+  exists fdb; split=> //.
+  have [_ _ ->] := dead_code_fd_meta ok_fdb.
+  have /= [_ _ _ <-] := [elaborate check_fundef_meta check_fda'].
+  have [_ _ ->] := dead_code_fd_meta ok_fda.
+  done.
 Qed.
 
 (* TODO: move *)
@@ -319,16 +343,6 @@ Lemma compiler_third_part_alloc_ok entries (p p' : sprog) (fn: funname) (m: mem)
   alloc_ok p fn m.
 Proof. case/compiler_third_partP => _; exact. Qed.
 
-Lemma check_no_ptrP entries ao u fn :
-  check_no_ptr entries ao = ok u →
-  fn \in entries →
-  allNone (sao_params (ao fn)) ∧ allNone (sao_return (ao fn)).
-Proof.
-  clear.
-  case: u => /allMP h /InP ok_fn; move: (h _ ok_fn).
-  by t_xrbindP.
-Qed.
-
 Lemma allNone_nth {A} (m: seq (option A)) i :
   allNone m ->
   nth None m i = None.
@@ -336,6 +350,30 @@ Proof.
   elim: m i.
   - by move => ? _; exact: nth_nil.
   by case => // m ih [].
+Qed.
+
+Lemma check_wf_ptrP entries p ao fn fd :
+  check_wf_ptr entries p ao = ok tt ->
+  fn \in entries ->
+  get_fundef p.(p_funcs) fn = Some fd ->
+  all2 (λ (x : var_i) pi, wptr_status x == omap pp_writable pi)
+    (f_params fd) (sao_params (ao fn)) /\
+  let n := find (fun x => wptr_status x != Some true) fd.(f_params) in
+  sao_return (ao fn) = [seq Some i | i <- iota 0 n] ++ nseq (size (sao_return (ao fn)) - n) None.
+Proof.
+  move=> hcheck /InP ok_fn get_fd.
+  set n := find (fun x => wptr_status x != Some true) fd.(f_params).
+  move: hcheck; rewrite /check_wf_ptr.
+  t_xrbindP=> /allMP -/(_ _ ok_fn); rewrite get_fd => /assertP hparams.
+  move=> /allMP -/(_ _ ok_fn); rewrite get_fd -/n => /assertP /andP [/eqP hl hr].
+  split=> //.
+  rewrite -{1}(cat_take_drop n (sao_return _)).
+  apply f_equal2 => //.
+  apply (@eq_from_nth _ None).
+  + by rewrite size_drop size_nseq.
+  move=> i; rewrite size_drop => hi.
+  rewrite nth_nseq hi.
+  exact: allNone_nth hr.
 Qed.
 
 Lemma sem_call_length {dc:DirectCall}(p: uprog) scs m fn va scs' m' vr :
@@ -355,6 +393,57 @@ Proof.
   by exists fd.
 Qed.
 
+Lemma keep_only_false {A} (l : seq A) n :
+  size l <= n ->
+  keep_only l (nseq n false) = [::].
+Proof. by elim: n l => [|n ih] [|a l] //= /ih. Qed.
+
+Lemma keep_only_true {A} (l : seq A) n :
+  keep_only l (nseq n true) = l.
+Proof. by elim: n l => [//|n ih] [//|a l] /=; rewrite ih. Qed.
+
+Lemma keep_only_cat {A} (l1 l2 : seq A) (tokeep1 tokeep2 : seq bool) :
+  size l1 = size tokeep1 ->
+  keep_only (l1 ++ l2) (tokeep1 ++ tokeep2) =
+  keep_only l1 tokeep1 ++ keep_only l2 tokeep2.
+Proof. by elim: tokeep1 l1 => [|b tokeep1 ih] [|a1 l1] //= [] /ih ->; case: b. Qed.
+
+Definition value_uincl_or_in_mem {A} (m : mem) (o : option A) (v v' : value) :=
+  match o with
+  | None => value_uincl v v'
+  | Some _ => value_in_mem m v v'
+  end.
+
+Definition size_glob (p:sprog) := Z.of_nat (size (sp_globs (p_extra p))).
+Definition get_wptrs (p:uprog) fn :=
+  oapp (fun fd => map wptr_status fd.(f_params)) [::] (get_fundef p.(p_funcs) fn).
+Definition get_align_args (p:sprog) fn :=
+  oapp (fun fd => fd.(f_extra).(sf_align_args)) [::] (get_fundef p.(p_funcs) fn).
+Definition get_nb_wptr (p:uprog) fn :=
+  find (fun wptr => wptr != Some true) (get_wptrs p fn).
+
+Lemma value_eq_or_in_mem_any_option {A B} (os1 : seq (option A)) (os2 : seq (option B)) m vs vs' :
+  List.Forall2 (fun o1 o2 => isSome o1 = isSome o2) os1 os2 ->
+  Forall3 (value_eq_or_in_mem m) os1 vs vs' ->
+  Forall3 (value_eq_or_in_mem m) os2 vs vs'.
+Proof.
+  move=> heq heqinmem.
+  elim: {os1 os2} heq vs vs' heqinmem.
+  move=> [|??] [|??] /List_Forall3_inv //=.
+  + by move=> _; constructor.
+  move=> o1 o2 os1 os2 heq _ ih [|??] [|??] /List_Forall3_inv // [heqinmem /ih{}ih].
+  constructor=> //.
+  by case: o1 o2 heq heqinmem => [?|] [?|].
+Qed.
+
+Lemma value_uincl_value_in_mem_trans {m} v2 v1 v3 :
+  value_uincl v1 v2 -> value_in_mem m v2 v3 -> value_in_mem m v1 v3.
+Proof.
+  move=> huincl [p [-> hread]].
+  exists p; split; first by reflexivity.
+  by move=> off w /(value_uincl_get_val_byte huincl); apply hread.
+Qed.
+
 Lemma compiler_front_endP
   entries
   (p: prog)
@@ -364,57 +453,267 @@ Lemma compiler_front_endP
   compiler_front_end aparams cparams entries p = ok p' →
   fn \in entries →
   sem_call (dc:=indirect_c) (wsw:= nosubword) p tt scs m fn va scs' m' vr →
-  extend_mem_eq m mi gd (sp_globs (p_extra p')) →
+  extend_mem m mi gd (sp_globs (p_extra p')) →
+  forall va',
+  wf_args (size_glob p') gd m mi (get_wptrs p fn) (get_align_args p' fn) va va' ->
+  Forall3 (value_eq_or_in_mem mi) (get_wptrs p fn) va va' ->
   alloc_ok p' fn mi →
-  ∃ vr' mi',
-    [/\
-     List.Forall2 value_uincl vr vr',
-     sem_call (dc:=direct_c) p' gd scs mi fn va scs' mi' vr' &
-     extend_mem_eq m' mi' gd (sp_globs (p_extra p'))
-    ].
+  ∃ vr' mi', [/\
+    sem_call (dc:=direct_c) p' gd scs mi fn va' scs' mi' vr',
+    extend_mem m' mi' gd (sp_globs (p_extra p')),
+    let n := get_nb_wptr p fn in
+      List.Forall2 (value_in_mem mi') (take n vr) (take n va') /\
+      List.Forall2 value_uincl (drop n vr) vr' &
+    mem_unchanged_params m mi mi' (get_wptrs p fn) va va'
+  ].
 Proof.
   rewrite /compiler_front_end;
-  t_xrbindP => p1 ok_p1 /check_no_ptrP checked_entries p2 ok_p2 p3.
+  t_xrbindP => p1 ok_p1 check_p1 p2 ok_p2 p3.
   rewrite print_sprogP => ok_p3 <- {p'} ok_fn exec_p.
-  rewrite (compiler_third_part_meta ok_p3) => m_mi ok_mi.
-  assert (ok_mi' : alloc_ok (sip := sip_of_asm_e) p2 fn mi).
-  - exact: compiler_third_part_alloc_ok ok_p3 ok_mi.
-  have := compiler_first_partP ok_p1 ok_fn exec_p.
-  case => {p ok_p1 exec_p} vr1 vr_vr1 exec_p1.
+  rewrite /size_glob (compiler_third_part_meta ok_p3) -/(size_glob _)
+    => m_mi va' va'_wf va'_eqinmem ok_mi.
+  have ok_mi': [elaborate alloc_ok p2 fn mi].
+  + exact: compiler_third_part_alloc_ok ok_p3 ok_mi.
+  have [vr1 vr_vr1 exec_p1] := compiler_first_partP ok_p1 ok_fn exec_p.
   have gd2 := sp_globs_stack_alloc ok_p2.
   rewrite -gd2 in ok_p2.
-  case/sem_call_length: (exec_p1) => fd [] ok_fd size_params size_tyin size_tyout size_res.
+  case/sem_call_length: (exec_p1) => fd1 [] get_fd1 size_params size_tyin size_tyout size_res.
   have! [mglob ok_mglob] := (alloc_prog_get_fundef ok_p2).
-  move=> /(_ _ _ ok_fd)[] fd' /alloc_fd_checked_sao[] ok_sao_p ok_sao_r ok_fd'.
-  move: checked_entries => /(_ _ ok_fn) [] params_noptr return_noptr.
-  assert (ok_va : wf_args (sp_globs (p_extra p2)) gd (ao_stack_alloc (stackalloc cparams p1)) m mi fn va va).
-  - move: params_noptr ok_sao_p.
-    rewrite size_params /wf_args.
-    move: (sao_params _); clear.
-    elim: va.
-    + by case => // _ _; constructor.
-    by move => v va ih [] // [] // pa /= /ih{}ih /succn_inj /ih{}ih; constructor.
-  have disjoint_va : disjoint_values (sao_params (ao_stack_alloc (stackalloc cparams p1) fn)) va va.
-  - rewrite /disjoint_values => i1 pi1 w1 i2 pi2 w2.
-    by rewrite (allNone_nth _ params_noptr).
+  move=> /(_ _ _ get_fd1)[] fd2 /dup[] ok_fd2 /alloc_fd_checked_sao[] ok_sao_p ok_sao_r get_fd2.
+  have [fd [get_fd _]] := sem_callE exec_p.
+  rewrite /get_nb_wptr /get_wptrs get_fd /= seq.find_map /preim.
+  set n := find _ _.
+  have := check_wf_ptrP check_p1 ok_fn get_fd.
+  rewrite -/n /= => -[check_params check_return].
+  move: check_params; rewrite all2_map -eqseq_all => /eqP check_params.
+  rewrite check_params.
+
+  have hargs: [elaborate
+    wf_args (size_glob p2) gd m mi
+      (map (omap pp_writable) (sao_params (ao_stack_alloc (stackalloc cparams p1) fn)))
+      (map (oapp pp_align U8) (sao_params (ao_stack_alloc (stackalloc cparams p1) fn)))
+      va va'].
+  + move: va'_wf; rewrite /get_wptrs get_fd /= check_params.
+    have [fd3 [get_fd3 align_eq]] := compiler_third_part_invariants ok_p3 get_fd2.
+    rewrite /get_align_args get_fd3 /= -align_eq.
+    move: ok_fd2; rewrite /alloc_fd.
+    by t_xrbindP=> _ _ <- /=.
+
+  have heqinmem:
+    Forall3 (value_eq_or_in_mem mi)
+      (sao_params (ao_stack_alloc (stackalloc cparams p1) fn)) va va'.
+  + move: va'_eqinmem; rewrite /get_wptrs get_fd /=.
+    apply: value_eq_or_in_mem_any_option.
+    rewrite check_params.
+    have /Forall2_flip :=
+      map_Forall2 (omap pp_writable)
+        (sao_params (ao_stack_alloc (stackalloc cparams p1) fn)).
+    apply Forall2_impl.
+    by move=> _ ? <-; apply isSome_omap.
+
   have := alloc_progP _ (hap_hsap haparams) ok_p2 exec_p1 m_mi.
-  move => /(_ (hap_hshp haparams) va ok_va disjoint_va ok_mi').
-  case => mi' [] vr2 [] exec_p2 [] m'_mi' [] ok_vr2 ?.
+  move => /(_ (hap_hshp haparams) va' hargs heqinmem ok_mi').
+  case => mi' [] vr2 [] exec_p2 m'_mi' vr2_wf vr2_eqinmem U.
   have [] := compiler_third_partP ok_p3.
-  case/(_ _ _ _ _ _ _ _ _ ok_fn exec_p2) => vr3 vr2_vr3 exec_p3.
-  exists vr3, mi'; split.
-  - apply: (Forall2_trans value_uincl_trans); first exact vr_vr1.
-    apply: (Forall2_trans value_uincl_trans); last exact vr2_vr3.
-    suff -> : vr1 = vr2 by exact: List_Forall2_refl.
-    move: ok_sao_r ok_vr2 return_noptr.
-    rewrite /wf_results size_res.
-    move: (sao_return _); clear.
-    elim: vr1 vr2.
-    + by case => // ??[] // _ /List_Forall3_inv.
-    move => r vr ih vr2 [] // [] // ns /= /succn_inj /ih{}ih /List_Forall3_inv.
-    by case: vr2 => // r2 vr2 [] /= -> /ih{}ih /ih ->.
-  - exact: exec_p3.
-  exact: m'_mi'.
+  case/(_ _ _ _ _ _ _ _ _ exec_p2).
+  set rminfo := fun fn => _.
+  move=> /= vr3 vr2_vr3 exec_p3 _.
+  exists vr3, mi'; split=> //.
+
+  have hle1: n <= size fd.(f_params) by apply find_size.
+  have [/esym size_vr1 /esym size_vr2] := Forall3_size vr2_wf.
+  have size_vr := Forall2_size vr_vr1.
+  have [/esym size_va /esym size_va'] := Forall3_size heqinmem.
+  have /(f_equal size) := check_params; rewrite 2!size_map => /esym size_sao_params.
+  have hle2: n <= size vr.
+  + have /(f_equal size) := check_return.
+    rewrite size_cat size_map size_iota -size_vr1 -size_vr => ->.
+    exact: leq_addr.
+
+  (* [vr2_eqinmem] can be split into two thanks to [check_results]:
+     - the first [n] elements satisfy [value_in_mem];
+     - the other ones satisfy equality. *)
+  have [vr2_inmem vr2_eq]:
+    List.Forall2 (value_in_mem mi') (take n vr1) (take n vr2) /\ drop n vr1 = drop n vr2.
+  + split.
+    + apply (nth_Forall2 (Vbool true) (Vbool true)).
+      + rewrite size_takel; last by rewrite -size_vr.
+        rewrite size_takel; last by rewrite size_vr2 -size_vr1 -size_vr.
+        reflexivity.
+      rewrite size_takel; last by rewrite -size_vr.
+      move=> i hi.
+      rewrite nth_take // nth_take //.
+      have := Forall3_nth vr2_eqinmem None (Vbool true) (Vbool true).
+      rewrite -size_vr1 -size_vr => /(_ _ (leq_trans hi hle2)).
+      rewrite check_return nth_cat size_map size_iota hi (nth_map 0);
+        last by rewrite size_iota.
+      by rewrite nth_iota.
+    apply (eq_from_nth (x0 := Vbool true)).
+    + by rewrite 2!size_drop size_vr1 size_vr2.
+    move=> i; rewrite size_drop -size_vr ltn_subRL => hi.
+    rewrite 2!nth_drop.
+    have := Forall3_nth vr2_eqinmem None (Vbool true) (Vbool true).
+    rewrite -size_vr1 -size_vr => /(_ _ hi).
+    rewrite check_return nth_cat size_map size_iota lt_nm_n.
+    by rewrite nth_nseq (ltn_sub2rE _ (leq_addr _ _)) -size_vr1 -size_vr hi.
+
+  (* [vr2_wf] can be rewritten into an equality thanks to [check_results] *)
+  have {}vr2_wf: take n vr2 = take n va'.
+  + apply (eq_from_nth (x0 := Vbool true)).
+    + rewrite size_takel; last by rewrite size_vr2 -size_vr1 -size_vr.
+      rewrite size_takel; last by rewrite size_va' size_sao_params.
+      reflexivity.
+    rewrite size_takel; last by rewrite size_vr2 -size_vr1 -size_vr.
+    move=> i hi.
+    rewrite nth_take // nth_take //.
+    have := Forall3_nth vr2_wf None (Vbool true) (Vbool true).
+    rewrite -size_vr1 -size_vr => /(_ _ (leq_trans hi hle2)).
+    rewrite check_return nth_cat size_map size_iota hi (nth_map 0);
+      last by rewrite size_iota.
+    by rewrite nth_iota //; case.
+
+  (* [fn_keep_only rminfo fn] is just [drop] thanks to [check_results] *)
+  have rminfo_vr2: fn_keep_only rminfo fn vr2 = drop n vr2.
+  + rewrite /fn_keep_only /rminfo ok_fn.
+    set k := size (sao_return (ao_stack_alloc (stackalloc cparams p1) fn)).
+    have ->:
+      [seq match i with
+           | Some _ => false
+           | None => true
+           end
+         | i <- sao_return (ao_stack_alloc (stackalloc cparams p1) fn)]
+      = nseq n false ++ nseq (k - n) true.
+    + rewrite check_return map_cat.
+      apply f_equal2.
+      + by rewrite -map_comp map_const_nseq size_iota.
+      by apply map_nseq.
+    case heq: all.
+    + by case: (n) heq => [|//] _; rewrite drop0.
+    rewrite -{1}(cat_take_drop n vr2).
+    rewrite keep_only_cat; last first.
+    + rewrite size_takel; last by rewrite size_vr2 -size_vr1 -size_vr.
+      by rewrite size_nseq.
+    rewrite keep_only_false; last first.
+    + by rewrite size_take; apply geq_minl.
+    by rewrite keep_only_true.
+
+  split.
+  + rewrite -vr2_wf.
+    by apply (Forall2_trans value_uincl_value_in_mem_trans (Forall2_take vr_vr1 n) vr2_inmem).
+  apply: (Forall2_trans value_uincl_trans); first exact (Forall2_drop vr_vr1 n).
+  by rewrite vr2_eq -rminfo_vr2.
+Qed.
+
+Lemma ptr_eq_wf_args glob_size gd m mi wptrs aligns va1 va2 va' :
+  Forall3 (fun o v v' => o <> None -> v = v') wptrs va1 va2 ->
+  wf_args glob_size gd m mi wptrs aligns va1 va' ->
+  wf_args glob_size gd m mi wptrs aligns va2 va'.
+Proof.
+  move=> heqs hargs.
+  move=> i; move: (hargs i); rewrite /wf_arg.
+  case ok_writable: nth => [writable|//].
+  move=> [p [-> hargp]].
+  exists p; split; first by reflexivity.
+  have hi := nth_not_default ok_writable ltac:(discriminate).
+  have := Forall3_nth heqs None (Vbool true) (Vbool true) hi;
+    rewrite ok_writable => /(_ ltac:(discriminate)) <-.
+  case: hargp => halign hover hvalid hfresh hwnglob hdisj.
+  split=> //.
+  move=> hw j vaj pj neq_ij /isSomeP [writablej ok_writablej] ok_vaj ok_pj.
+  apply: (hdisj hw _ _ _ neq_ij _ _ ok_pj).
+  + by rewrite ok_writablej.
+  have hj := nth_not_default ok_writablej ltac:(discriminate).
+  have := Forall3_nth heqs None (Vbool true) (Vbool true) hj;
+    rewrite ok_writablej => /(_ ltac:(discriminate)) ->.
+  done.
+Qed.
+
+Lemma ptr_eq_mem_unchanged_params m1 m2 m3 wptrs vs1 vs2 vs' :
+  Forall3 (fun o v v' => o <> None -> v = v') wptrs vs1 vs2 ->
+  mem_unchanged_params m1 m2 m3 wptrs vs2 vs' ->
+  mem_unchanged_params m1 m2 m3 wptrs vs1 vs'.
+Proof.
+  move=> heqs hunch p hvalid hnvalid hdisj.
+  apply (hunch p hvalid hnvalid).
+  elim: {wptrs vs1 vs'} hdisj vs2 heqs {hunch}.
+  + move=> [|??] /List_Forall3_inv // _.
+    by constructor.
+  move=> wptr v1 v' wptrs vs1 vs' hdisj _ ih [|v2 vs2]
+    /List_Forall3_inv //= [heq /ih{}ih].
+  constructor=> //.
+  case: wptr heq hdisj => [writable|//].
+  by move=> /(_ ltac:(discriminate)) <-.
+Qed.
+
+(* [compiler_front_endP] takes [value_eq_or_in_mem] as an hypothesis. But to call
+   the lemma in the context of lemma [compile_prog_to_asmP], we rather need
+   [value_uincl_or_in_mem] as an hypothesis. *)
+Lemma compiler_front_endP_uincl
+  entries
+  (p: prog)
+  (p': @sprog _pd _ _asmop)
+  (gd : pointer)
+  scs m mi fn va scs' m' vr :
+  compiler_front_end aparams cparams entries p = ok p' →
+  fn \in entries →
+  sem_call (dc:=indirect_c) (wsw:= nosubword) p tt scs m fn va scs' m' vr →
+  extend_mem m mi gd (sp_globs (p_extra p')) →
+  forall va',
+  wf_args (size_glob p') gd m mi (get_wptrs p fn) (get_align_args p' fn) va va' ->
+  Forall3 (value_uincl_or_in_mem mi) (get_wptrs p fn) va va' ->
+  alloc_ok p' fn mi →
+  ∃ vr' mi', [/\
+    sem_call (dc:=direct_c) p' gd scs mi fn va' scs' mi' vr',
+    extend_mem m' mi' gd (sp_globs (p_extra p')),
+    let n := get_nb_wptr p fn in
+      List.Forall2 (value_in_mem mi') (take n vr) (take n va') /\
+      List.Forall2 value_uincl (drop n vr) vr' &
+    mem_unchanged_params m mi mi' (get_wptrs p fn) va va'
+  ].
+Proof.
+  move=> ok_p' ok_fn exec_p m_mi va' va'_wf va'_uinmem ok_mi.
+
+  (* We define [va2] as [va] where non-reg ptr values are replaced by values
+     from [va']. We can prove
+     [Forall3 (value_eq_or_in_mem mi) (get_wptrs p fn) va2 va'] and thus
+     call [compiler_front_endP] on [va2] and [va']. *)
+  set va2 :=
+    map3 (fun o v v' => if o is Some _ then v else v') (get_wptrs p fn) va va'.
+  have [size_va size_va'] := Forall3_size va'_uinmem.
+
+  have huincl: List.Forall2 value_uincl va va2.
+  + elim: {va va' exec_p va'_wf size_va size_va'} va'_uinmem @va2 => /=.
+    + by constructor.
+    move=> wptr v v' wptrs va va' huinmem _ ih.
+    constructor=> //.
+    by case: wptr huinmem => [writable|] /=.
+  have [vr2 [exec2_p huincl2]] := psem.sem_call_uincl huincl exec_p.
+
+  have hptreq: Forall3 (fun o v v' => o <> None -> v = v') (get_wptrs p fn) va va2.
+  + elim: {va va' exec_p va'_wf size_va size_va'} va'_uinmem @va2 {huincl exec2_p} => /=.
+    + by constructor.
+    move=> wptr v v' wptrs va va' _ _ ih.
+    constructor=> //.
+    by case: wptr.
+  have hargs := ptr_eq_wf_args hptreq va'_wf.
+
+  have heqinmem: Forall3 (value_eq_or_in_mem mi) (get_wptrs p fn) va2 va'.
+  + elim: {va va' exec_p va'_wf size_va size_va'} va'_uinmem @va2
+      {huincl exec2_p hptreq hargs} => /=.
+    + by constructor.
+    move=> wptr v v' wptrs va va' huinmem _ ih.
+    constructor=> //.
+    by case: wptr huinmem => [writable|] /=.
+
+  have [vr' [mi' [p'_call m'_mi' vr2_vr' U]]] :=
+    compiler_front_endP ok_p' ok_fn exec2_p m_mi hargs heqinmem ok_mi.
+
+  exists vr', mi'; split=> //.
+  + case: vr2_vr' => hres1 hres2.
+    split.
+    + by apply: Forall2_trans value_uincl_value_in_mem_trans (Forall2_take huincl2 _) hres1.
+    by apply: Forall2_trans value_uincl_trans (Forall2_drop huincl2 _) hres2.
+  by apply: (ptr_eq_mem_unchanged_params hptreq U).
 Qed.
 
 Lemma compiler_back_end_meta entries (p: sprog) (tp: lprog) :
@@ -618,8 +917,18 @@ Proof.
     + by move=> <-.
     move=> hmm.
     split.
-    + move=> pr w ok_w.
-      have := M'.(read_incl) ok_w.
+    + move=> /= pr hb hval.
+      have := M'.(read_incl_mem) hb hval.
+      rewrite hmm.(read_untouched) //.
+      apply not_between_U8_disjoint_zrange => //.
+      move=> /(zbetween_trans bottom_instack).
+      rewrite -/(between _ _ _ _) -pointer_range_between => /pointer_rangeP.
+      have ss := sem_call_stack_stable_sprog exec_p.
+      rewrite ss.(ss_limit) (ss_top_stack ss).
+      have := [elaborate top_stack_below_root _ m']; rewrite -/(top_stack _) /=.
+      by Lia.lia.
+    + move=> pr w hb ok_w.
+      have := M'.(read_incl_stk) hb ok_w.
       rewrite hmm.(read_untouched) //.
       apply not_between_U8_disjoint_zrange => //.
       move=> /(zbetween_trans bottom_instack).
@@ -822,7 +1131,7 @@ Qed.
 *)
 
 Record mem_agreement_with_ghost (m m': mem) (gd: pointer) (data: seq u8) (ma_ghost: mem) : Prop :=
-  { ma_extend_mem_eq : extend_mem_eq m ma_ghost gd data
+  { ma_extend_mem : extend_mem m ma_ghost gd data
   ; ma_match_mem : match_mem ma_ghost m'
   ; ma_stack_stable : stack_stable m ma_ghost
   ; ma_stack_range : (wunsigned (stack_limit ma_ghost) <= wunsigned (top_stack m'))%Z
@@ -831,6 +1140,51 @@ Record mem_agreement_with_ghost (m m': mem) (gd: pointer) (data: seq u8) (ma_gho
 Definition mem_agreement (m m': mem) (gd: pointer) (data: seq u8) : Prop :=
   ∃ ma_ghost, mem_agreement_with_ghost m m' gd data ma_ghost.
 
+Definition get_asm_align_args (xp:asm_prog) fn :=
+  oapp (fun xd => xd.(asm_fd_align_args)) [::] (get_fundef xp.(asm_funcs) fn).
+
+Lemma compiler_back_end_to_asm_get_fundef entries sp xp fn :
+  compiler_back_end_to_asm aparams cparams entries sp = ok xp ->
+  fn \in entries ->
+  exists sfd xd, [/\
+    get_fundef sp.(p_funcs) fn = Some sfd,
+    get_fundef xp.(asm_funcs) fn = Some xd,
+    xd.(asm_fd_export) &
+    sfd.(f_extra).(sf_align_args) = xd.(asm_fd_align_args)].
+Proof.
+  move=> ok_xp ok_fn.
+  move: ok_xp; rewrite /compiler_back_end_to_asm.
+  t_xrbindP=> lp ok_lp ok_xp.
+  move: ok_lp; rewrite /compiler_back_end.
+  t_xrbindP=> hcheck _ lp1 ok_lp1 lp2.
+  rewrite print_linearP => ok_lp2 lp3.
+  rewrite print_linearP => ok_lp3.
+  rewrite print_linearP => ?; subst lp.
+
+  have [sfd [get_sfd ranone]]:
+    exists sfd,
+      get_fundef sp.(p_funcs) fn = Some sfd
+      /\ is_RAnone sfd.(f_extra).(sf_return_address).
+  + move: hcheck; rewrite /check_export.
+    have /InP ok_fn' := ok_fn.
+    move=> /allMP -/(_ fn ok_fn').
+    case: get_fundef => //.
+    t_xrbindP=> sfd ranone.
+    by exists sfd.
+
+  have get_lfd1 := [elaborate get_fundef_p' ok_lp1 get_sfd].
+  have [lfd2 ok_lfd2 get_lfd2] := [elaborate
+    stack_zeroization_lprog_get_fundef ok_lp2 get_lfd1].
+  have get_lfd3 := [elaborate get_fundef_tunnel_program ok_lp3 get_lfd2].
+  have [xd get_xd ok_xd] := [elaborate ok_get_fundef ok_xp get_lfd3].
+
+  exists sfd, xd.
+  move/assemble_fdI: ok_xd => [_ _ [_ [_ [_ [_ _ _ {-2}-> _]]]]] /=.
+  move/stack_zeroization_lfd_invariants: ok_lfd2 => [_ _ _ _ _ _ <- _ _ [_ <-]] /=.
+  by split.
+Qed.
+
+(* FIXME: do we want to ensure wf_result_pointer? the only thing that matters is value_mem_uincl, it seems *)
 Lemma compile_prog_to_asmP
   entries
   (p : prog)
@@ -843,7 +1197,8 @@ Lemma compile_prog_to_asmP
   compile_prog_to_asm aparams cparams entries p = ok xp
   -> fn \in entries
   -> psem.sem_call (dc:= indirect_c) (wsw:=nosubword) p tt scs m fn va scs' m' vr
-  -> mem_agreement m (asm_mem xm) (asm_rip xm) (asm_globs xp)
+  -> forall mi,
+     mem_agreement_with_ghost m (asm_mem xm) (asm_rip xm) (asm_globs xp) mi
   -> enough_stack_space xp fn (top_stack m) (asm_mem xm)
   -> exists xd : asm_fundef,
       [/\
@@ -851,36 +1206,67 @@ Lemma compile_prog_to_asmP
         , asm_fd_export xd
         &   asm_scs xm = scs
             -> asm_reg xm ad_rsp = top_stack m
-            -> List.Forall2 value_uincl va (get_typed_reg_values xm (asm_fd_arg xd))
+            -> wf_args (Z.of_nat (size (asm_globs xp))) (asm_rip xm) m mi
+                (get_wptrs p fn) (get_asm_align_args xp fn) va (get_typed_reg_values xm (asm_fd_arg xd))
+            -> Forall3 (value_uincl_or_in_mem (asm_mem xm)) (get_wptrs p fn) va (get_typed_reg_values xm (asm_fd_arg xd))
             -> exists xm',
                 [/\ asmsem_exportcall xp fn xm xm'
                   , mem_agreement m' (asm_mem xm') (asm_rip xm') (asm_globs xp), asm_scs xm' = scs'
                   , (cparams.(stack_zero_info) fn <> None ->
-                      forall p, ~ validw m Aligned p U8 ->
-                        read xm'.(asm_mem) Aligned p U8 = read xm.(asm_mem) Aligned p U8
-                        \/ read xm'.(asm_mem) Aligned p U8 = ok 0%R)
-                  & List.Forall2 value_uincl vr (get_typed_reg_values xm' (asm_fd_res xd))
+                      forall pr, ~ validw m Aligned pr U8 ->
+                        Forall3 (disjoint_from_writable_param (ep:=ep_of_asm_e) pr) (get_wptrs p fn) va (get_typed_reg_values xm (asm_fd_arg xd)) ->
+                        read xm'.(asm_mem) Aligned pr U8 = read xm.(asm_mem) Aligned pr U8
+                        \/ read xm'.(asm_mem) Aligned pr U8 = ok 0%R)
+                  & let n := get_nb_wptr p fn in
+                    List.Forall2 (value_in_mem (asm_mem xm')) (take n vr) (take n (get_typed_reg_values xm (asm_fd_arg xd))) /\
+                    List.Forall2 value_uincl (drop n vr) (get_typed_reg_values xm' (asm_fd_res xd))
                 ]
       ].
 Proof.
-  rewrite /compile_prog_to_asm; t_xrbindP => sp ok_sp ok_xp ok_fn p_call [] mi.
+  rewrite /compile_prog_to_asm; t_xrbindP => sp ok_sp ok_xp ok_fn p_call mi.
   have -> := compiler_back_end_to_asm_meta ok_xp.
   case=> /= mi1 mi2 mi3 mi4.
   rewrite (ss_top_stack mi3).
   move=> /dup[] henough /(enough_stack_space_alloc_ok ok_xp ok_fn mi4) ok_mi.
-  have := compiler_front_endP ok_sp ok_fn p_call mi1 ok_mi.
-  case => vr' [] mi' [] vr_vr' sp_call m1.
+  have [sfd [xd [get_sfd get_xd xd_export align_args_eq]]] :=
+    compiler_back_end_to_asm_get_fundef ok_xp ok_fn.
+  exists xd; split=> //.
+  move=> ok_scs ok_rsp va_wf va_uinmem.
+  have hargs: [elaborate
+    wf_args (size_glob sp) (asm_rip xm) m mi (get_wptrs p fn)
+      (get_align_args sp fn) va (get_typed_reg_values xm (asm_fd_arg xd))].
+  + rewrite /get_align_args get_sfd /=.
+    move: va_wf.
+    rewrite /get_asm_align_args get_xd /=.
+    by rewrite align_args_eq.
+  have huinmem:
+    Forall3 (value_uincl_or_in_mem mi) (get_wptrs p fn) va
+      (get_typed_reg_values xm (asm_fd_arg xd)).
+  + have [hsize1 hsize2] := Forall3_size va_uinmem.
+    apply (nth_Forall3 None (Vbool true) (Vbool true)) => // i hi.
+    have := Forall3_nth va_uinmem None (Vbool true) (Vbool true) hi.
+    case ok_writable: nth => [writable|//].
+    move=> [pr [ok_pr hread]]; rewrite ok_pr.
+    exists pr; split; first by reflexivity.
+    move=> off w /[dup] /get_val_byte_bound hoff /hread ok_w.
+    move: (va_wf i); rewrite /wf_arg ok_writable ok_pr.
+    move=> [_ [[<-] hargp]].
+    rewrite -ok_w; apply (match_mem_read_incl_mem mi2).
+    apply hargp.(wap_valid).
+    by apply (between_byte hargp.(wap_no_overflow) (zbetween_refl _ _) hoff).
+
+  have := compiler_front_endP_uincl ok_sp ok_fn p_call mi1 hargs huinmem
+    (enough_stack_space_alloc_ok ok_xp ok_fn mi4 henough).
+  case => vr' [] mi' [] sp_call m1 vr_vr' U.
   have := compiler_back_end_to_asmP ok_xp ok_fn sp_call.
-  case => xd [] ok_xd export /(_ _ _ erefl _ mi2) xp_call.
-  exists xd; split => //.
-  move=> ok_scs ok_rsp va_args'.
+  rewrite get_xd.
+  case => _ [] [<-] _ /(_ _ _ erefl _ mi2) xp_call.
   have hallocatable: allocatable_stack mi (asm_fd_total_stack xd).
   + rewrite /allocatable_stack.
-    by have /= := henough _ ok_xd; Lia.lia.
-  have := xp_call ok_scs ok_rsp va_args' hallocatable.
+    by have /= := henough _ get_xd; Lia.lia.
+  have := xp_call ok_scs ok_rsp (List_Forall2_refl _ value_uincl_refl) hallocatable.
   case => xm' [] {} xp_call m2 ok_scs' hzero vr'_res'.
-  exists xm'; split => //;
-    last exact: Forall2_trans value_uincl_trans vr_vr' vr'_res'.
+  exists xm'; split => //.
   + case: xp_call => _ _ _ /= _ /asmsem_invariantP /= xm_xm' _.
     exists mi'; split.
     - rewrite -(asmsem_invariant_rip xm_xm').
@@ -894,24 +1280,26 @@ Proof.
       -(ss_limit (sem_call_stack_stable_sprog sp_call))
       -(ss_top_stack (asmsem_invariant_stack_stable xm_xm')).
     exact: mi4.
-  move=> hszs pr hnvalid.
-  have := mi1.(eme_valid) pr.
-  case: (@idP (between _ _ _ _)) => [hb|_].
-  + move=> _; left.
-    have [i [hi ->]]:
-      exists i,
-        (0 <= i < (Z.of_nat (size (sp_globs (p_extra sp)))))%Z
-        /\ (pr = asm_rip xm + wrepr _ i)%R.
-    + exists (wunsigned pr - wunsigned (asm_rip xm))%Z; split; last first.
-      + rewrite wrepr_sub !wrepr_unsigned.
-        by rewrite GRing.addrC GRing.subrK.
-      move: hb; rewrite /between /zbetween wsize8 !zify /=.
-      by Lia.lia.
-    have /mi2.(read_incl) -> := mi1.(eme_read_new) hi.
-    by have /m2.(read_incl) -> := m1.(eme_read_new) hi.
-  rewrite orbF => hvalideq.
-  apply (hzero hszs pr).
-  by rewrite -hvalideq.
+  + move=> hszs pr hnvalid hdisj.
+    change reg_size with Uptr in pr.
+    case: (@idP (validw mi Aligned pr U8)).
+    + move=> hvalid.
+      left.
+      have := U _ hvalid hnvalid hdisj.
+      rewrite (match_mem_read_incl_mem mi2 hvalid).
+      rewrite (sem_call_validw_stable_sprog sp_call) in hvalid.
+      by rewrite (match_mem_read_incl_mem m2 hvalid).
+    move=> hnvalid'.
+    by apply hzero.
+  split; last first.
+  + by apply: (Forall2_trans value_uincl_trans (proj2 vr_vr')).
+  move: vr_vr'.
+  move=> /= [vr_vr' _].
+  apply: Forall2_impl vr_vr'.
+  move=> v1 _ [pr [-> hread]].
+  exists pr; split; first by reflexivity.
+  move=> off w /hread.
+  by apply (mm_read_ok m2).
 Qed.
 
 End PROOF.